+++ @remirror/core-extensions/src/constants.ts
@@ -1,31 +1,3 @@
-import {
-  BaseKeymapExtension,
-  CompositionExtension,
-  HistoryExtension,
-  NodeCursorExtension,
-  PlaceholderExtension,
-  SSRHelperExtension,
-} from './extensions';
-import {
-  BoldExtension,
-  CodeExtension,
-  ItalicExtension,
-  LinkExtension,
-  StrikeExtension,
-  UnderlineExtension,
-} from './marks';
-import {
-  BlockquoteExtension,
-  BulletListExtension,
-  CodeBlockExtension,
-  HardBreakExtension,
-  HeadingExtension,
-  HorizontalRuleExtension,
-  ImageExtension,
-  ListItemExtension,
-  OrderedListExtension,
-} from './nodes';
-
 /**
  * Used to denote a node is empty.
  *
@@ -39,35 +11,3 @@ export const EMPTY_NODE_CLASS_SELECTOR = `.${EMPTY_NODE_CLASS_NAME}`;
  * arrow key presses between items in general. Used by both the CompositionExtension and NodeCursorExtension
  */
 export const NODE_CURSOR_DEFAULTS = ['emoji'];
-
-export const ALL_NODE_EXTENSIONS = [
-  BlockquoteExtension,
-  BulletListExtension,
-  CodeBlockExtension,
-  HardBreakExtension,
-  HeadingExtension,
-  HorizontalRuleExtension,
-  ImageExtension,
-  ListItemExtension,
-  OrderedListExtension,
-];
-
-export const ALL_MARK_EXTENSIONS = [
-  BoldExtension,
-  CodeExtension,
-  ItalicExtension,
-  LinkExtension,
-  StrikeExtension,
-  UnderlineExtension,
-];
-
-export const ALL_GENERIC_EXTENSIONS = [
-  BaseKeymapExtension,
-  CompositionExtension,
-  HistoryExtension,
-  NodeCursorExtension,
-  PlaceholderExtension,
-  SSRHelperExtension,
-];
-
-export const ALL_EXTENSIONS = [...ALL_NODE_EXTENSIONS, ...ALL_MARK_EXTENSIONS, ...ALL_GENERIC_EXTENSIONS];
